Create Calculator enhanced
I have added the functionality of repeatingly performing calculations(getting numbers)  until user press exit.For the first time it asks for 2 nos,and next time onwards it will ask for only one number.Moreover,I have added a  facility for displaying previous output.
for eg if user gives first 2 nos as 23,24 and wants to perform addition it will show 23+24=47. if he enters 3rd number as 33 and wants to perform multiplication it will show 47*33=1551 and so on.